pyproject.toml
[tool.poetry]
name = "nginx-limitreq-ipset"
version = "0.1.0"
description = ""
authors = ["Sven Ulland <sven.ulland@gmail.com>"]

[tool.poetry.scripts]
nginx-limitreq-ipset = "nginx_limitreq_ipset.cli:cli"

[tool.poetry.dependencies]
python = "^3.10"
python-json-logger = "^2.0.2"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = ">=6.2.4"
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
